requirements.txt
hydra-core
numpy
torch
click==7.1.2 # Spacy breaks for click>=8.0
spacy
allennlp
git+https://github.com/eric-mitchell/higher@master # For in-place functional models
git+https://github.com/eric-mitchell/transformers@master # To enable gradient disabling for some models
datasets
jsonlines
wandb
sentence-transformers
